+++ src/main/java/meep/task/ToDo.java
@@ -27,7 +27,6 @@ public class ToDo extends Task {
 
     @Override
     public String toString() {
-
         return "[ T ]" + super.toString();
     }
 }
