+++ src/main/java/meep/textui/Meep.java
@@ -1,4 +1,4 @@
-package meep.testui;
+package meep.textui;
 
 import java.io.IOException;
